feat: split schemas and API handlers for parkings and place types into dedicated modules
