+++ apps/admin-client/src/features/parkings/schemas/index.ts
@@ -1,44 +1,2 @@
-import { z } from 'zod';
-
-export const parkingListBaseInputSchema = z.object({
-  page: z.coerce.number().int().positive().optional(),
-  limit: z.coerce.number().int().positive().max(100).optional(),
-  search: z.string().max(255).optional(),
-});
-
-export const parkingListItemSchema = z.object({
-  id: z.uuid(),
-  name: z.string(),
-  active: z.boolean(),
-  place: z.object({
-    id: z.uuid(),
-    name: z.string(),
-  }),
-  organization: z.object({
-    id: z.uuid(),
-    name: z.string(),
-  }),
-  version: z.number().int().positive(),
-});
-
-export type ParkingListItemSchema = z.infer<typeof parkingListItemSchema>;
-
-export const parkingListSchema = z.object({
-  data: z.array(parkingListItemSchema),
-  total: z.number().int().nonnegative(),
-  currentPage: z.number().int().positive(),
-});
-
-export const placeTypesListItemSchema = z.object({
-  id: z.uuid(),
-  name: z.string(),
-  version: z.number().int().positive(),
-});
-
-export type PlaceTypesListItemSchema = z.infer<typeof placeTypesListItemSchema>;
-
-export const placeTypesListSchema = z.object({
-  data: z.array(placeTypesListItemSchema),
-  total: z.number().int().nonnegative(),
-  currentPage: z.number().int().positive(),
-});
+export * from './place-types.ts';
+export * from './parking.ts';
